Initialize the ChromeBrowserProxyResolver only once.
The ChromeBrowserProxyResolver instance was being initialized on every
update check, making it register a listener for the proxy resolved
signal on each update, calling the callback multiple times.

Only the first time the callback is called will actually do something,
and the rest of the calls will only log an error message and return.

This patch removes the error message by registering the callback only once.

Bug: None
TEST=emerge-link update_engine; deployed on a link and ran several update checks.

Change-Id: I8a64fdf168966a895080760b8a95f9d8bf27976c