Implement a new resetStatus() method in Android interface.
When an alredy applied update is deleted from the server (normally
because is was detected to be a bad update), we need to go back to
the idle state and remove the update to prevent it from breaking
more devices.

This patch allows the application side to reset the applied update
back to idle.

Bug: 27081760
TEST=Deployed on a non-Brillo device, sent resetStatus.

(cherry picked from commit 3b678db2e7fd4baa38c6d6b8bb8334f7d88b8682)

Change-Id: I654c904f4b4f606445a70ec95a660491391a779f