Don't rebuild .img files on incremental builds.
The sample image files are extracted from a .tar.bz2 file. The images
were generated at some point in the past and this date is stored in the
tarball file, so the rule that extracts them from the tarball keeps
this date, making the .img file have an older date than the file that
generates it (the .tar.bz2). This casues the build system to rebuild
those .img files every time unnecesarily.

This patch updates the date on the extracted .img while extracting it
fixing this problem.

Bug: None
TEST=Running `mmma system/update_engine` twice doesn't regenerate the .img files.

Change-Id: I76dc536c5ee45898ad5c43fd005fcf30e4311cd1