brillo_update_payload: Keep the partitions order.
The order in associative array is based on hash, but we need to make sure
it's always rootfs and then kernel for Chrome OS.

This patch adds a PARTITIONS_ORDER array and pass it to extract_image()
for target image. For Android/Brillo, it will use ab_partitions_list.txt
if found, for Chrome OS, it's always ( "system" "boot" ).

Bug: None
Test: Generated payload for edison and peppy.

Change-Id: I29672befa14045573cba497784b67629a21578bf