Use provided omaha_url and app_version on non-interactive update.
The D-Bus API allows to start a non-interactive update check and pass
the Omaha URL and the app_version to be used. If updates from other
Omaha servers are allowed (like in non-official builds or dev-mode with
debug options enabled) these provided options would normally be used,
except for non-interactive update.

This patch uses the provided url and app_version even for
non-interactive updates.

Bug: 26937101
TEST=Ran `update_engine_client --nointeractive --check_for_update --omaha_url=http://127.0.0.1:1234` on edison.

Change-Id: I8a1e7fdc6ede434537ee028ef59e3eee29a44edc