payload_generator: Mark all zero blocks in the old partition as visited.
When processing duplicated and zeroed blocks, discard all the zeroed
blocks from the source partition since it doesn't make sense to waste
I/O reading them, and all the new zeroed blocks are taken care of in a
different way.

This also solves the problem that those zeroed blocks might actually not
be written in the source slot when flashing an image with fastboot.

Bug: 28626303
TEST=deployed an update on a device after flashing it with fastboot.

Change-Id: I42012a84634a5b35f181b6b22e8a738ba3b70957