Generate SOURCE_COPY and ZERO operations beyond the end of the filesystem.
This patch analyzes the whole partition size, regardless of what the
size of the filesystem in the partition is. This makes the blocks in
the ext2 partitions past the end of the filesystem (normally verity
hashes) be compared against the old blocks and produce SOURCE_COPY
operations for those as well if possible.

Bug: 29272878
TEST=Generated a payload to itself of a build with verity hashes, only
SOURCE_COPY (and REPLACE_BZ of zeroed blocks) are produced.

Change-Id: If7f2733da41c333d4a8247b1473ad06fb63466d6