Improve logs when the update fails due to source hash mistmatch.
The source hash mismatch at the operation level is the most likelly
error when applying an incremental update on top of a modified image.
This error can also be caught when re-reading the whole target partition
at the end of the update, but if a discrepancy on any operation is
found when applying the operations, it will fail earlier.

This patch improves the log message when this situation occurs. This is
not a common error case so a more verbose error message is acceptable
here. Previously, the error message was not giving any clue about why
was it failing.

Bug: 29163155
Bug: 30510360
TEST=Modified a partition and applied a payload. Error message is nicer.

(cherry picked from commit 6714084915ea7b66accf933344b0d4cd042dff3d)

Change-Id: I0e7ad33006b0d2475bbb3681ad313fb3eb4b3747