Initialize the CertificateChecker on Android.
The CertificateChecker singleton is a class that keeps track of the
SSL certificates seen for the download and updatecheck servers. In
Android devices, we use the download server only, but the class was not
initialized leading to CURLE_FAILED_INIT failure and a log message,
except that DLOG messages are not shown on "release" builds and
therefore the message was lost.

This patch initializes the certificate checker class, allowing
update_engine to download HTTPS files.

Bug: 29585834
TEST=Tested on a device that update_engine can download a payload
directly from an https:// URL.

Change-Id: I70ec15aea620802bb52ca8405b9444d5b88d0288