Reset the update progress when partition verification fails.
When the update completes and there is an error verifying the new
partitions we need to reset the update progress proactivelly to
prevent the update failing repeatedly until we hit the 10 retries
limit. This patch will reset the update progress when the payload
application fails to verify the contents of the newly written
partition.

Bug: 30162558
TEST=Push an update, cancel it at 90%, erase some blocks in system,
continue the update (restarts at 90% and then fails), launch the update
again (restarts from 0%).

(cherry picked from commit 5990bf33b50beec0caf55a97f5ca87608ccbc694)

Change-Id: Ie0b22bb190397aa342ca09d55c2b9dfd6ad5f239