Implement a memory-based Prefs class.
When the /data storage is not available, we need to store the required
prefs in memory to keep all the code working. This patch splits the
Prefs class functionality and implements a MemoryPrefs class which
stores all the values in memory.

Bug: 27178350
TEST=Added unittest for the MemoryPrefs.

(cherry picked from commit a0284ace424d1653f36208da24b5ce878d393c7e)

Change-Id: Iac6002b5c9b6c6f2cfd2d02f0a2aa6a4a8970404