DO NOT MERGE: Parse Android .map files for unknown filesystem.
When generating a filesystem during the Android build, we also generate
a text file with .map extension with the list of files and the blocks
in the filesystem they are located.

For filesystems unsupported in delta_generator (like squashfs) we use
this text file to produce efficient delta payloads.

Bug: 28150981
Test: Added unittest for parsing. Generated a delta payload of a squashfs image.

(cherry-pick from 20bdc70e540e021c4a26553dfdbfd2fabe8ea432)

Change-Id: I5b83241622903259d5d40ae73f3f908756a77afd