Don't clear previous partitions in DeltaPerformer.
When we used to have a FilesystemVerifierAction before DownloadAction,
partitions in InstallPlan was filled in by FilesystemVerifierAction
with source hashes, but now that action is gone, DeltaPerformer is the
first action that touches partitions, so clearing partitions are not
needed. In multi payload case, partitions in previous payloads are
stored there so it should not be removed.

We could also move the partitions into Payload but that will make our
code more complex (keeping track of both payload index and partition
index in FilesystemVerifierAction and PostinstallRunnerAction) thus
more likely to have bugs.

This simple patch solves the same problem without polluting the code.

Bug: 36252799
Test: FilesystemVerifierAction verifies all partitions in all payloads.
Change-Id: I2aed02389bf047a6dedf59b306434ccea4eebca8
(cherry picked from commit cf6bd59a3cd149bd86d763e96ede81ce102d64d9)