Treat noupdate="true" as noupdate.
In order to support cases where not all <app> have an update, we
still need to send a delta to self payload because update_engine
doesn't know which partition should be copied over and we need the
hash in payload to detect any data corruption.

But Omaha doesn't allow noupdate updatecheck status with a payload,
so we introduced a new attribute in postinstall action to tell
update_engine this is a delta to self payload and can be treated as
no update if all <app> set it.

Bug: 36252799
Test: mma
Change-Id: I377f2dccacf7b92c5bce91ceacb7752ffbdbc480
(cherry picked from commit 1b70c3a7f257d3ff166f33bbac761ad9b7e969e1)