update_engine: Reduce verbosity of delta_generator
This outputs ~4000 lines of the form:

  0.00%         48 REPLACE_BZ <boot-operation-2>

which is not very useful in the build. Update it to show only unique lines
with a count of how many times that line was repeated.

BUG=chromium:737687
TEST=FEATURES=test emerge-reef update_engine
See the output. I have not added a specific test for this since it is
really just debug output IMO.
Change-Id: I15149733216ba0516c62f99e5968768ec4fa3d20
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/553579
Reviewed-by: Ben Chan <benchan@chromium.org>